codigobot/requirements.txt
discord.py
discord-components
PyNaCl
googletrans
googletrans==4.0.0-rc1